requirements.txt

black
pillow

